refactor: draft out Problem and PythonProblem classes
